requirements.txt
fastapi
uvicorn
spacy
pydantic
openai
requests
python-dotenv
chromadb
sqlalchemy
